% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{all_means}
\alias{all_means}
\title{All Vowel Means}
\format{\preformatted{Classes ‘rowwise_df’, ‘tbl_df’ and 'data.frame':	9254 obs. of  7 variables:
 $ idstring  : chr  "PH00-1-1-" "PH00-1-1-" "PH00-1-1-" "PH00-1-1-" ...
 $ DOB       : int  1979 1979 1979 1979 1979 1979 1979 1979 1979 1979 ...
 $ sex       : chr  "m" "m" "m" "m" ...
 $ plt_vclass: chr  "*hr" "ae" "aeBR" "aeh" ...
 $ F1_n      : num  -0.2865 0.8021 0.2339 -0.0761 1.6408 ...
 $ F2_n      : num  -0.246 0.404 0.643 0.953 -1.034 ...
 $ n         : int  61 220 22 48 2 43 155 395 93 165 ...
}}
\source{
Philadelphia Neighborhood Corpus
}
\usage{
all_means
}
\description{
Vowel means from the PNC
}
\keyword{datasets}

