% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ar1_summary_df}
\alias{ar1_summary_df}
\title{Stan model summary}
\format{\preformatted{'data.frame':	9727 obs. of  11 variables:
 $ mean   : num  -1.17e-01 3.19e-01 1.86e-01 -7.94e-05 -3.06e-03 ...
 $ se_mean: num  0.00147 0.00187 0.00187 0.00215 0.00194 ...
 $ sd     : num  0.147 0.187 0.187 0.215 0.194 ...
 $ 2.5\%   : num  -0.406 -0.052 -0.174 -0.43 -0.389 ...
 $ 25\%    : num  -0.215 0.19 0.055 -0.143 -0.132 ...
 $ 50\%    : num  -0.11699 0.32075 0.18667 0.00136 0.00101 ...
 $ 75\%    : num  -0.0147 0.4444 0.3125 0.1451 0.1266 ...
 $ 97.5\%  : num  0.169 0.688 0.554 0.422 0.378 ...
 $ n_eff  : num  10000 10000 10000 10000 10000 10000 10000 10000 10000 10000 ...
 $ Rhat   : num  1 1 1 1 1 ...
 $ param  : chr  "speaker_effects[1,1]" "speaker_effects[1,2]" "speaker_effects[1,3]" "speaker_effects[1,4]" ...
}}
\usage{
ar1_summary_df
}
\description{
Stan model summary
}
\keyword{datasets}

