% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ays_faith}
\alias{ays_faith}
\title{/ay/ before faithful /t/, /d/}
\format{\preformatted{'data.frame':	2819 obs. of  65 variables:
 $ vowel_index       : int  4 2 2 2 2 2 2 2 2 2 ...
 $ word_trans        : chr  "AO2 L R AY1 T" "S AY1 D" "S AY1 D" "D AY1 D" ...
 $ word              : chr  "alright" "side" "side" "died" ...
 $ age               : int  21 21 21 21 21 21 21 21 21 21 ...
 $ city              : chr  "NA" "NA" "NA" "NA" ...
 $ ethnicity         : chr  "i/r" "i/r" "i/r" "i/r" ...
 $ first_name        : chr  "NA" "NA" "NA" "NA" ...
 $ last_name         : chr  "NA" "NA" "NA" "NA" ...
 $ name              : chr  "NA" "NA" "NA" "NA" ...
 $ sex               : chr  "m" "m" "m" "m" ...
 $ state             : chr  "NA" "NA" "NA" "NA" ...
 $ tiernum           : int  0 0 0 0 0 0 0 0 0 0 ...
 $ year              : int  2000 2000 2000 2000 2000 2000 2000 2000 2000 2000 ...
 $ years_of_schooling: chr  "14" "14" "14" "14" ...
 $ vowel             : chr  "AY" "AY" "AY" "AY" ...
 $ stress            : int  1 1 1 1 1 1 1 1 1 1 ...
 $ F1                : num  631 624 666 673 541 ...
 $ F2                : num  1445 1403 1369 1373 1304 ...
 $ F3                : chr  "1868.4" "2387" "2271.9" "2420.9" ...
 $ B1                : num  89.4 45.7 48.9 81.3 71.5 ...
 $ B2                : num  181 220 156 158 97 ...
 $ B3                : chr  "95.7" "133.4" "617.2" "169.6" ...
 $ dur               : num  0.079 0.109 0.25 0.22 0.12 0.189 0.13 0.109 0.2 0.13 ...
 $ plt_vclass        : chr  "ay0" "ay" "ay" "ay" ...
 $ plt_manner        : chr  "stop" "stop" "stop" "stop" ...
 $ plt_place         : chr  "apical" "apical" "apical" "apical" ...
 $ plt_voice         : chr  "voiceless" "voiced" "voiced" "voiced" ...
 $ plt_preseg        : chr  "liquid" "oral_apical" "oral_apical" "oral_apical" ...
 $ plt_folseq        : chr  "" "" "" "" ...
 $ style             : chr  "NA" "NA" "NA" "NA" ...
 $ glide             : chr  "s" "s" "" "" ...
 $ pre_seg           : chr  "R" "S" "S" "D" ...
 $ fol_seg           : chr  "T" "D" "D" "D" ...
 $ context           : chr  "internal" "internal" "internal" "internal" ...
 $ F1_20_            : chr  "584.8" "567.5" "634.9" "572.9" ...
 $ F2_20_            : chr  "1406.5" "1365.7" "1416.1" "1593.9" ...
 $ F1_35_            : chr  "628.9" "601.9" "665.3" "653.7" ...
 $ F2_35_            : chr  "1433.2" "1389.1" "1367.4" "1437.2" ...
 $ F1_50_            : chr  "594.7" "623.8" "639.5" "717.9" ...
 $ F2_50_            : chr  "1508.2" "1386.2" "1386.4" "1380.7" ...
 $ F1_65_            : chr  "587.1" "609" "579.7" "695.7" ...
 $ F2_65_            : chr  "1609.2" "1471.2" "1640.9" "1410" ...
 $ F1_80_            : chr  "563.7" "547.9" "528.6" "586.5" ...
 $ F2_80_            : chr  "1707.8" "1599.4" "1800.1" "1622.4" ...
 $ nFormants         : int  6 5 5 5 5 4 6 6 5 4 ...
 $ F1_n              : num  0.866 0.777 1.296 1.381 -0.265 ...
 $ F2_n              : num  -0.148 -0.284 -0.392 -0.379 -0.6 ...
 $ F1_20_n           : chr  "0.284380153632963" "0.0685038858971636" "0.909547842278256" "0.135887229583482" ...
 $ F1_35_n           : chr  "0.834677460404568" "0.497760741972973" "1.28889111043827" "1.14414170548248" ...
 $ F1_50_n           : chr  "0.407916283724549" "0.771037635811933" "0.966948468381417" "1.94525479153094" ...
 $ F1_65_n           : chr  "0.313080466684544" "0.586357360523504" "0.220740329040329" "1.66823437859829" ...
 $ F1_80_n           : chr  "0.0210859773771619" "-0.176072694890217" "-0.416905756583912" "0.305593428497175" ...
 $ F2_20_n           : chr  "-0.271812670236099" "-0.402218603176084" "-0.241128921309044" "0.327159678610796" ...
 $ F2_35_n           : chr  "-0.186473493532726" "-0.327426965166387" "-0.396785022636918" "-0.173688598146453" ...
 $ F2_50_n           : chr  "0.053243294959895" "-0.336696014321435" "-0.336056769552121" "-0.35427524547756" ...
 $ F2_65_n           : chr  "0.376061903463291" "-0.0650169873631312" "0.477382199399505" "-0.26062588677311" ...
 $ F2_80_n           : chr  "0.691209574734923" "0.344738909766922" "0.986221035773174" "0.418252058237992" ...
 $ idstring          : chr  "PH00-1-1-" "PH00-1-1-" "PH00-1-1-" "PH00-1-1-" ...
 $ DOB               : int  1979 1979 1979 1979 1979 1979 1979 1979 1979 1979 ...
 $ pre_sp            : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
 $ fol_sp            : logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
 $ pre_misc          : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
 $ fol_misc          : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
 $ SUBTLWF           : num  79.3 200.9 200.9 157.2 171.4 ...
 $ word_status       : chr  "coda" "coda" "coda" "coda" ...
}}
\source{
The Philadelphia Neighborhood Corpus
}
\usage{
ays_faith
}
\description{
/ay/ before faithful /t/, /d/
}
\keyword{datasets}

