% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ays_flap}
\alias{ays_flap}
\title{/ay/ before flapped /t/, /d/}
\format{\preformatted{'data.frame':	580 obs. of  71 variables:
 $ word              : chr  "fighting" "decided" "decided" "decided" ...
 $ age               : int  21 21 52 52 18 18 18 18 18 18 ...
 $ city              : logi  NA NA NA NA NA NA ...
 $ ethnicity         : Factor w/ 27 levels "g","g/i","g/r",..: 7 7 12 12 NA NA NA NA NA NA ...
 $ first_name        : logi  NA NA NA NA NA NA ...
 $ last_name         : logi  NA NA NA NA NA NA ...
 $ name              : logi  NA NA NA NA NA NA ...
 $ sex               : Factor w/ 2 levels "f","m": 2 2 1 1 2 2 2 2 2 2 ...
 $ state             : logi  NA NA NA NA NA NA ...
 $ tiernum           : int  0 0 0 0 0 0 0 0 0 0 ...
 $ year              : int  2000 2000 2000 2000 2000 2000 2000 2000 2000 2000 ...
 $ years_of_schooling: Factor w/ 16 levels "10","11","12",..: 5 5 3 3 3 3 3 3 3 3 ...
 $ vowel             : Factor w/ 1 level "AY": 1 1 1 1 1 1 1 1 1 1 ...
 $ stress            : int  1 1 1 1 1 1 1 1 1 1 ...
 $ F1                : num  589 667 839 856 620 ...
 $ F2                : num  1130 1445 1513 1800 1394 ...
 $ F3                : num  1753 2418 2761 2807 2311 ...
 $ B1                : num  40.8 19.5 139.4 144.8 36.1 ...
 $ B2                : num  176.5 228.4 60.9 198.6 152 ...
 $ B3                : num  1759 202 646 571 271 ...
 $ dur               : num  0.07 0.07 0.16 0.12 0.089 0.2 0.06 0.13 0.09 0.1 ...
 $ plt_vclass        : Factor w/ 2 levels "ay","ay0": 2 1 1 1 1 1 1 2 2 2 ...
 $ plt_manner        : Factor w/ 1 level "stop": 1 1 1 1 1 1 1 1 1 1 ...
 $ plt_place         : Factor w/ 1 level "apical": 1 1 1 1 1 1 1 1 1 1 ...
 $ plt_voice         : Factor w/ 2 levels "voiced","voiceless": 2 1 1 1 1 1 1 2 2 2 ...
 $ plt_preseg        : Factor w/ 10 levels "?","liquid","nasal_apical",..: 7 6 6 6 6 2 6 7 2 3 ...
 $ plt_folseq        : Factor w/ 2 levels "one_fol_syll",..: 1 1 1 1 1 1 1 1 1 1 ...
 $ style             : Factor w/ 11 levels "?","C","G","K",..: NA NA NA NA 1 11 1 1 11 8 ...
 $ glide             : Factor w/ 3 levels "?","m","s": 3 3 3 2 2 3 2 1 1 2 ...
 $ pre_seg           : Factor w/ 14 levels "B","F","G","HH",..: 2 11 11 11 11 10 11 2 10 9 ...
 $ fol_seg           : Factor w/ 2 levels "D","T": 2 1 1 1 1 1 1 2 2 2 ...
 $ context           : Factor w/ 1 level "word_flap": 1 1 1 1 1 1 1 1 1 1 ...
 $ vowel_index       : int  2 4 4 4 4 2 4 2 2 4 ...
 $ word_trans        : Factor w/ 69 levels "AA0 R TH R AY1 T AH0 S",..: 24 16 16 16 16 54 16 22 58 69 ...
 $ F1_20_            : num  634 598 638 820 524 ...
 $ F2_20_            : num  1156 1480 1439 1786 1439 ...
 $ F1_35_            : num  560 600 766 803 552 ...
 $ F2_35_            : num  1017 1430 1480 1679 1398 ...
 $ F1_50_            : num  560 650 838 854 606 ...
 $ F2_50_            : num  1063 1452 1519 1812 1385 ...
 $ F1_65_            : num  601 682 749 767 611 ...
 $ F2_65_            : num  1183 1480 1650 1792 1406 ...
 $ F1_80_            : num  597 688 580 599 561 ...
 $ F2_80_            : num  1287 1551 1683 1748 1426 ...
 $ nFormants         : int  6 5 5 4 5 6 5 5 6 6 ...
 $ F1_n              : num  0.333 1.305 1.329 1.433 1.191 ...
 $ F2_n              : num  -1.157 -0.15 -0.468 0.192 -0.14 ...
 $ F1_20_n           : num  0.8921 0.4478 0.0701 1.2075 0.0469 ...
 $ F1_35_n           : num  -0.0288 0.4778 0.8728 1.103 0.3804 ...
 $ F1_50_n           : num  -0.0201 1.0955 1.3245 1.4209 1.0283 ...
 $ F1_65_n           : num  0.489 1.495 0.765 0.88 1.084 ...
 $ F1_80_n           : num  0.438 1.568 -0.291 -0.171 0.493 ...
 $ F2_20_n           : num  -1.0725 -0.0366 -0.6384 0.1584 0.0143 ...
 $ F2_35_n           : num  -1.5164 -0.1957 -0.5441 -0.0871 -0.125 ...
 $ F2_50_n           : num  -1.37 -0.126 -0.454 0.22 -0.171 ...
 $ F2_65_n           : num  -0.9865 -0.0366 -0.1529 0.1733 -0.1005 ...
 $ F2_80_n           : num  -0.6528 0.19 -0.077 0.0727 -0.029 ...
 $ DOB               : int  1979 1979 1948 1948 1982 1982 1982 1982 1982 1982 ...
 $ word_status       : Factor w/ 1 level "onset": 1 1 1 1 1 1 1 1 1 1 ...
 $ phrase_status     : Factor w/ 1 level "onset": 1 1 1 1 1 1 1 1 1 1 ...
 $ streetid          : Factor w/ 50 levels "PH00-1-","PH02-2-",..: 1 1 1 1 1 1 1 1 1 1 ...
 $ speakerid         : Factor w/ 197 levels "PH00-1-1-","PH00-1-2-",..: 1 1 2 2 3 3 3 3 3 3 ...
 $ flap_coding       : Factor w/ 8 levels "?","app","approx",..: 3 1 1 1 1 1 1 6 6 6 ...
 $ vocing_bar        : Factor w/ 3 levels "?","0","1": 3 1 1 1 1 1 1 3 3 3 ...
 $ formants          : Factor w/ 4 levels "?","0","1","o": 3 1 1 1 1 1 1 3 3 3 ...
 $ idstring          : chr  "PH00-1-1-" "PH00-1-1-" "PH00-1-2-" "PH00-1-2-" ...
 $ pre_sp            : logi  FALSE FALSE FALSE FALSE FALSE TRUE ...
 $ fol_sp            : logi  FALSE FALSE FALSE TRUE FALSE TRUE ...
 $ pre_misc          : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
 $ fol_misc          : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
 $ SUBTLWF           : num  70.8 88.7 88.7 88.7 88.7 ...
}}
\source{
The Philadelphia Neighborhood Corpus
}
\usage{
ays_flap
}
\description{
/ay/ before flapped /t/, /d/
}
\keyword{datasets}

