% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{flap_tracks}
\alias{flap_tracks}
\title{full flap tracks}
\format{\preformatted{Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	100980 obs. of  47 variables:
 $ id                : int  29 29 29 29 29 29 29 29 29 29 ...
 $ age               : int  21 21 21 21 21 21 21 21 21 21 ...
 $ ethnicity         : chr  "i/r" "i/r" "i/r" "i/r" ...
 $ sex               : chr  "m" "m" "m" "m" ...
 $ tiernum           : int  0 0 0 0 0 0 0 0 0 0 ...
 $ year              : int  2000 2000 2000 2000 2000 2000 2000 2000 2000 2000 ...
 $ years_of_schooling: num  14 14 14 14 14 14 14 14 14 14 ...
 $ vowel             : chr  "AY" "AY" "AY" "AY" ...
 $ stress            : int  1 1 1 1 1 1 1 1 1 1 ...
 $ word              : chr  "DIED" "DIED" "DIED" "DIED" ...
 $ F1_meas           : num  635 635 635 635 635 ...
 $ F2_meas           : num  1471 1471 1471 1471 1471 ...
 $ F3_meas           : num  2408 2408 2408 2408 2408 ...
 $ F1                : num  533 554 567 588 614 ...
 $ F2                : num  1661 1625 1567 1531 1516 ...
 $ F3                : num  2452 2171 1801 1872 2190 ...
 $ B1                : num  63.2 61 56.1 57.3 55.4 ...
 $ B2                : num  94.2 383.3 1079.8 1077.8 473.7 ...
 $ B3                : num  2452 2171 1801 1872 2190 ...
 $ dur               : num  0.189 0.189 0.189 0.189 0.189 0.189 0.189 0.189 0.189 0.189 ...
 $ plt_vclass        : chr  "ay" "ay" "ay" "ay" ...
 $ plt_manner        : chr  "stop" "stop" "stop" "stop" ...
 $ plt_place         : chr  "apical" "apical" "apical" "apical" ...
 $ plt_voice         : chr  "voiced" "voiced" "voiced" "voiced" ...
 $ plt_preseg        : chr  "oral_apical" "oral_apical" "oral_apical" "oral_apical" ...
 $ plt_folseq        : chr  "" "" "" "" ...
 $ pre_seg           : chr  "D" "D" "D" "D" ...
 $ fol_seg           : chr  "D" "D" "D" "D" ...
 $ context           : chr  "faithful" "faithful" "faithful" "faithful" ...
 $ vowel_index       : int  2 2 2 2 2 2 2 2 2 2 ...
 $ word_trans        : chr  "D AY1 D" "D AY1 D" "D AY1 D" "D AY1 D" ...
 $ t_rel             : num  0.019 0.028 0.038 0.047 0.057 ...
 $ t_prop            : num  0.1 0.147 0.2 0.247 0.3 ...
 $ percent           : num  0.1 0.15 0.2 0.25 0.3 0.35 0.4 0.45 0.5 0.55 ...
 $ idstring          : chr  "PH00-1-1-" "PH00-1-1-" "PH00-1-1-" "PH00-1-1-" ...
 $ F1_n              : num  -0.3579 -0.1062 0.0684 0.3276 0.6474 ...
 $ F2_n              : num  0.5417 0.4252 0.2419 0.1257 0.0795 ...
 $ word_status       : chr  "coda" "coda" "coda" "coda" ...
 $ phrase_status     : chr  "coda" "coda" "coda" "coda" ...
 $ word_fol_seg2     : chr  NA NA NA NA ...
 $ phrase_fol_seg2   : chr  NA NA NA NA ...
 $ speakerid         : chr  NA NA NA NA ...
 $ n                 : int  NA NA NA NA NA NA NA NA NA NA ...
 $ flapid            : int  NA NA NA NA NA NA NA NA NA NA ...
 $ flap_coding       : Factor w/ 8 levels "?","app","approx",..: NA NA NA NA NA NA NA NA NA NA ...
 $ fol_sp            : logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
 $ t0                : num  20.7 20.7 20.8 20.8 20.8 ...
}}
\source{
The Philadelphia Neighborhood Corpus
}
\usage{
flap_tracks
}
\description{
full flap tracks
}
\keyword{datasets}

