% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{mid_low_means}
\alias{mid_low_means}
\title{Means for mid and low vowels}
\format{A data frame with 652 rows}
\source{
Philadelphia Neighborhood Corpus
}
\usage{
mid_low_means
}
\description{
Means for mid and low vowels
}
\keyword{datasets}

