% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{simulation}
\alias{simulation}
\title{Simulation
A simulation of categorical phonetic change}
\format{\preformatted{Classes ‘grouped_df’, ‘tbl_df’, ‘tbl’ and 'data.frame':	11000 obs. of  6 variables:
 $ p          : num  0 0 0 0 0 0 0 0 0 0 ...
 $ sim_id     : int  1 12 23 34 45 56 67 78 89 100 ...
 $ n          : int  60 60 60 60 60 60 60 60 60 60 ...
 $ F1_mean    : num  -0.010253 0.025661 0.013791 -0.011533 -0.000701 ...
 $ F1_sd      : num  0.43 0.529 0.502 0.441 0.407 ...
 $ F1_kurtosis: num  2.93 6.21 6.05 2.79 3.2 ...
 - attr(*, "vars")=List of 1
  ..$ : symbol p
 - attr(*, "drop")= logi TRUE
}}
\source{
see \code{system.file(run_sim/run_sim.R, package = "phoneticChangne")}
}
\usage{
simulation
}
\description{
Simulation
A simulation of categorical phonetic change
}
\keyword{datasets}

