% Generated by roxygen2 (4.0.2): do not edit by hand
\name{syllabify}
\alias{syllabify}
\title{Syllabify}
\usage{
syllabify(trans)
}
\description{
Syllabify
}
\examples{
syllabify(c("S", "T", "R", "EH1", "NG", "TH", "AH0", "N"))
syllabify(c("M", "AE2", "S", "T", "ER0"))
syllabify(c("B","AE","N","T","ER"))
syllabify(c("P","R","AE","NG","K","S","T","ER","Z"))
syllabify("S AH0 L AE1 B IH0 F AH0 K EY2 SH AH0 N")
}

