% Generated by roxygen2 (4.0.2): do not edit by hand
\name{verboten_string}
\alias{verboten_string}
\title{Create Verboten String}
\usage{
verboten_string(x = c("PH76-3-1-", "PH84-2-5-", "PH84-2-3-", "PH92-1-1-",
  "PH82-1-16-", "PH80-2-5-", "PH80-2-4-", "PH85-1-5-", "PH84-1-2-", "PH84-1-1-",
  "PH77-3-4-"))
}
\description{
Create Verboten String
}

