% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{white}
\alias{white}
\title{Basic Demographic Info from the PNC}
\format{A data frame with 326 rows
\describe{
  \item{sex}{speaker sex}
  \item{year}{year of interview}
  \item{ethnicity}{single character code for reported ethnicity}
  \item{schooling}{highest educational attainment}
  \item{transcribed}{how many seconds of transcribed speech}
  \item{total}{total duration of recording}
  \item{nvowels}{number of vowels measured}
  \item{idstring}{unique idstring for a speaker}
}}
\source{
Philadelphia Neighborhood Corpus
}
\usage{
white
}
\description{
Basic Demographic Info from the PNC
}
\keyword{datasets}

